requirements.txt
selenium
openai
pillow
pyperclip
python-dotenv
pydantic
